💄 style: Fix rerender
