+++ package.json
@@ -116,6 +116,7 @@
     "emoji-mart": "^5.6.0",
     "fast-deep-equal": "^3.1.3",
     "framer-motion": "^12.23.11",
+    "harden-react-markdown": "^1.0.4",
     "immer": "^10.1.1",
     "katex": "^0.16.22",
     "leva": "^0.10.0",
@@ -147,6 +148,7 @@
     "remark-github": "^12.0.0",
     "remark-math": "^6.0.0",
     "shiki": "^3.8.1",
+    "streamdown": "^1.0.11",
     "swr": "^2.3.4",
     "ts-md5": "^1.3.1",
     "unified": "^11.0.5",
